[Gold II] Title: 도토리 숨기기, Time: 4 ms, Memory: 2632 KB -BaekjoonHub

+++ 백준/Gold/15732. 도토리 숨기기/도토리 숨기기.cc
@@ -4,18 +4,18 @@ using namespace std;
 
 struct rule
 {
-	long long s,e,j;
+	int s,e,j;
 	rule(int s,int e, int j):s(s),e(e),j(j){};
 };
 
 
 int main() {
     ios_base::sync_with_stdio(false);cin.tie(NULL); cout.tie(NULL);
-    long long result;
-    long long n, k;
-	long long d;
+    int result;
+    int n, k;
+	int d;
     cin>>n>>k>>d;
-    long long s=0,&e=n;
+    int s=0,&e=n;
     vector <rule> rule;
     for(int i=0; i<k; i++){
         int a,b,c;
@@ -25,7 +25,7 @@ int main() {
     while (s<=e)
     {
         long long cnt=0;
-        long long m = (s+e)/2;
+        int m = (s+e)/2;
         for(auto&r : rule){
             if(m < r.s) continue;
 			cnt += (min(m,r.e)-r.s)/r.j + 1;
